Create  SK Doom, Complete + Revised
This is a short-story revolving around three main characters; the Doctor, his pet robot, and a villain who tries to take over his lab (and eventually the world). I originally wrote this story as a series of small chapters in July 2004 when I was ten and distributed said content on a child-oriented forum which catered to a small community. I revised the grammar and spelling in August 2011, after deciding the material was worth disseminating to a larger audience that could appreciate its worth. The original was in much needed revision, so here is the complete, remastered version. Enjoy